refactor: domain 패키지 생성
도메인 로직을 관리하기 위해 domain 패키지를 생성
- Lotto와 Lottos를 domain으로 이동
