+++ src/main/java/lotto/Application.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.domain.Lottos;
+
 public class Application {
 
     public static void main(String[] args) {
